requirements.txt
pandas
numpy
sentence-transformers
torch
nltk
selenium
matplotlib
seaborn
requests
python-dotenv